feat(runtime-c-api) Extract the `export` module.
